Implement OCCUPIED state in xworkspaces
For this, the m_clientlist is turned from a vector of window IDs into a
map, associating each window with its desktop index.

The rebuild_clientlist() routine is mostly rewritten so that it updates
all of those desktop indices, in order to account for clients being
moved between desktops. Correspondingly, any changes to _NET_WM_DESKTOP
now also act as a trigger for a rebuild of the client list and desktop
states.

Since the desktop states are now directly influenced by the client list,
basically every call to rebuild_clientlist() has to be followed by a
subsequent call to rebuild_desktop_states().